# Get rid of vite-plugin-babel-macros