--- a/package.json
+++ b/package.json
@@ -28,7 +28,6 @@
     "@vitejs/plugin-react": "^1.0.7",
     "babel-plugin-macros": "^3.1.0",
     "typescript": "^4.5.4",
-    "vite": "^2.8.0",
-    "vite-plugin-babel-macros": "^1.0.6"
+    "vite": "^2.8.0"
   }
 }