--- a/vite.config.ts
+++ b/vite.config.ts
@@ -1,8 +1,14 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-import macrosPlugin from "vite-plugin-babel-macros";
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [macrosPlugin(), react()],
-});
+  plugins: [
+    react({
+      babel: {
+        // Use .babelrc files, necessary to use LinguiJS CLI
+        babelrc: true,
+      },
+    }),
+  ],
+})