Getting rid of vite-plugin-babel-macros as suggested by @CamilleHbp

https://github.com/lingui/js-lingui/issues/1016#issuecomment-1102123689

Unfortunately this doesn't seem to change the performance a lot.

### with vite-plugin-babel-macros:

  Benchmark 1: rm -rf dist node_modules/.vite && yarn vite build

    Time (mean ± σ):     980.9 ms ±  44.4 ms    [User: 1281.7 ms, System: 108.1 ms]
    Range (min … max):   950.5 ms … 1070.8 ms    10 runs


### without vite-plugin-babel-macros (this PR):

  Benchmark 2: rm -rf dist node_modules/.vite && yarn vite build
  
    Time (mean ± σ):     965.0 ms ±  81.5 ms    [User: 1242.1 ms, System: 103.8 ms]
    Range (min … max):   926.9 ms … 1194.4 ms    10 runs